# Dryft/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Drift API
  description: Backend API for the Drift social/dating application.
  version: 1.0.0
  contact:
    name: Drift Team
servers:
  - url: https://api.drift.app
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /v1/users/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Auth]
      summary: Update current user profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                bio:
                  type: string
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Age Verification
  # ---------------------------------------------------------------------------
...
  /v1/matches/{matchID}:
    get:
      tags: [Matches]
      summary: Get a specific match
      operationId: getMatch
      parameters:
        - name: matchID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Match details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchWithUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Matches]
      summary: Unmatch a user
      operationId: unmatch
      parameters:
        - name: matchID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Unmatched successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get a user record (admin)
      operationId: getAdminUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/profile:
    get:
      tags: [Profile]
      summary: Get own profile
      operationId: getProfile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      tags: [Profile]
      summary: Update own profile
      operationId: updateOwnProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /v1/profile/preferences:
    get:
      tags: [Profile]
      summary: Get matching preferences
      operationId: getMatchingPreferences
      responses:
        "200":
          description: Matching preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchingPreferences"
    put:
      tags: [Profile]
      summary: Update matching preferences
      operationId: updateMatchingPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchingPreferences"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchingPreferences"
  /v1/profile/location:
    put:
      tags: [Profile]
      summary: Update location
      operationId: updateLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                  format: double
                longitude:
                  type: number
                  format: double
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  city:
                    type: string
                  country:
                    type: string
...
  /v1/users/{userId}:
    get:
      tags: [Profile]
      summary: Get a user's public profile
      operationId: getUserProfile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublicProfile"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Safety
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        bio:
          type: string
        profile_photo:
          type: string
        verified:
          type: boolean
        verified_at:
          type: string
          format: date-time
          nullable: true
        preferences:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MatchWithUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_a_id:
          type: string
          format: uuid
        user_b_id:
          type: string
          format: uuid
        matched_at:
          type: string
          format: date-time
        unmatched_at:
          type: string
          format: date-time
          nullable: true
        other_user:
          $ref: "#/components/schemas/User"
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        bio:
          type: string
        age:
          type: integer
        photos:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        verified:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserPublicProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        bio:
          type: string
        age:
          type: integer
        profile_photo:
          type: string
        photos:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        verified:
          type: boolean
    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
        bio:
          type: string
        interests:
          type: array
          items:
            type: string
...
    MatchingPreferences:
      type: object
      properties:
        min_age:
          type: integer
          minimum: 18
        max_age:
          type: integer
          maximum: 100
        max_distance_km:
          type: integer
        gender_preference:
          type: string
          enum: [male, female, all]
        show_me_to:
          type: string
          enum: [male, female, all]

    # Safety schemas
...